General code refactoring and improvements
