--- src/features/auth/Auth.module.css
@@ -1,26 +1,0 @@
-.registerForm{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    gap: 0.8rem;
-    padding: 0.8rem 1rem;
-    margin:auto;
-    max-width: 500px;
-
-}
-
-.textField {
-    width: 100%;
-  }
-  
-.input {
-padding: 10px 14px;
-font-size: 0.9rem;
-}
-
-.input::placeholder {
-font-size: 0.8rem;
-color: gray;
-}
