+++ src/features/auth/components/UserRegistrationStep.tsx
@@ -1,10 +1,11 @@
-import styles from "../Auth.module.css";
+import styles from "../../../shared/styles/ParkingForm.module.css";
 import { Box, Link, Typography } from "@mui/material"
 import AuthFormContainer from "./AuthFormContainer"
-import { FieldsType, FormUserValues, UserRegistrationStepProps } from "../types";
+import { FormUserValues, UserRegistrationStepProps } from "../types";
 import InputForm from "./InputForm";
 import { grey } from "@mui/material/colors";
 import ButtonPrimary from "../../../shared/ui/components/ButtonPrimary";
+import { FieldsType } from "../../../shared/types";
 
 
 const fields: FieldsType[] = [
