+++ src/features/auth/pages/LoginPage.tsx
@@ -1,4 +1,4 @@
-import styles from "../Auth.module.css";
+import styles from "../../../shared/styles/ParkingForm.module.css";
 import { Box, Link, Typography } from "@mui/material";
 import { Link as RouterLink, useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
@@ -11,6 +11,7 @@ import { FormValues } from "../types";
 import HeaderForm from "../components/HeaderForm";
 import authService from "../services/AuthService";
 import { useAuthStore, User } from "../../../store/auth.store";
+import { showError, showSuccess } from "../../../shared/ui/toast";
 
 const LoginPage = () => {
   const navigate = useNavigate();
@@ -24,16 +25,19 @@ const LoginPage = () => {
     resolver: yupResolver(loginSchema),
   });
   const onSubmit = async (data: FormValues) => {
-    const response = await authService.login(data.email, data.password)
-    if (response.token) {
-      login(response.token, response.user )
-      reset();
-      alert("Estacionamiento registrado");
-      navigate("/");
-    } else {
-      //podria devolver usuario inexistemte
-      //contrasenña invalida
+    try {
+      const response = await authService.login(data.email, data.password)
+      if (response.token) {
+        login(response.token, response.user )
+        reset();
+        navigate("/");
+        showSuccess(`Bienvenido ${response.user.email}`);
+      } 
+    } catch (err) {
+      showError('Hubo un error al registrar el usuario');
     }
+    const response = await authService.login(data.email, data.password)
+   
     
     
   };
